--- src/Arrays.java
@@ -1,16 +1,0 @@
-public class Arrays {
-    public static int[] generateRandomArray() {
-        java.util.Random random = new java.util.Random();
-        int[] arr = new int[30];
-        for (int i = 0; i < arr.length; i++) {
-            arr[i] = random.nextInt(100_000) + 100_000;
-        }
-        return arr;
-    }
-
-    public static void main(String[] args) {
-        int[] wages = generateRandomArray();
-
-    }
-}
-
